Revert type of resource values back to int

+++ OpenRA.Mods.Common/MapGenerator/Terraformer.cs
@@ -2110,7 +2110,7 @@ namespace OpenRA.Mods.Common.MapGenerator
 				var oldValue = CheckValue3By3(cpos);
 				Map.Resources[mpos] = new ResourceTile(
 					resourceType.ResourceIndex,
-					resourceType.MaxDensity);
+					(byte)resourceType.MaxDensity);
 				var newValue = CheckValue3By3(cpos);
 				return newValue - oldValue;
 			}
